Fixed readResponseStream not parsing gzipped content when in http error
When the HTTP server responded with an error (statusCode >= 400),
the readResponseStream didn't decode the inStream as gzip even if
the request sended had the Content-Encoding header.

My fix moves the code that parses Content-Encoded header and
decodes gzipped body out the test to get the input stream to the
point where we are certain that we have a inStream that we can
decode.